experiments/nofiltergpt-chat/requirements.txt
colorama==0.4.6
python-dotenv==1.1.1
requests==2.32.4
